requirements.txt
pandas===1.0.3
xlsxwriter===1.2.8
xlrd==1.2.0
dns.resolver
dns.reversename